Fixed compiler warnings after uClibc code update

+++ src/libm/e_atan2.c
@@ -57,8 +57,8 @@ double attribute_hidden __ieee754_atan2(double y, double x)
 	ix = hx&0x7fffffff;
 	EXTRACT_WORDS(hy,ly,y);
 	iy = hy&0x7fffffff;
-	if(((ix|((lx|-lx)>>31))>0x7ff00000)||
-	   ((iy|((ly|-ly)>>31))>0x7ff00000))	/* x or y is NaN */
+	if(((ix|((lx|-(int32_t)lx)>>31))>0x7ff00000)||
+	   ((iy|((ly|-(int32_t)ly)>>31))>0x7ff00000))	/* x or y is NaN */
 	   return x+y;
 	if(((hx-0x3ff00000)|lx)==0) return atan(y);   /* x=1.0 */
 	m = ((hy>>31)&1)|((hx>>30)&2);	/* 2*sign(x)+sign(y) */
